backend/requirements.txt
fastapi==0.104.1
uvicorn==0.40.0
transformers>=4.35.0
torch>=2.9.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.5.0
cachetools>=5.3.0
aioredis>=2.0.1
gunicorn>=21.2.0
asyncio>=3.4.3
windows-curses>=2.3.2